Move PlayersValueBuffer to separate module
