+++ tools/twix/src/main.rs
@@ -37,6 +37,7 @@ mod panel;
 mod panels;
 mod twix_painter;
 mod value_buffer;
+mod players_value_buffer;
 
 fn setup_logger() -> Result<(), InitError> {
     Dispatch::new()
